backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
boto3==1.38.36
    # via -r requirements.in
botocore==1.38.36
    # via
    #   boto3
    #   s3transfer
build==1.2.2.post1
    # via pip-tools
chardet==5.2.0
    # via reportlab
click==8.2.1
    # via pip-tools
cssselect2==0.8.0
    # via svglib
dj-database-url==3.0.0
    # via -r requirements.in
django==5.2.3
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-ckeditor
    #   django-classy-tags
    #   django-cms
    #   django-cors-headers
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-js-asset
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-storage-url
    #   django-storages
    #   django-treebeard
    #   django-vite
    #   djangocms-admin-style
    #   djangocms-alias
    #   djangorestframework
    #   easy-thumbnails
django-ckeditor==6.7.3
    # via -r requirements.in
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==5.0.1
    # via
    #   -r requirements.in
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-frontend
    #   djangocms-link
django-cors-headers==4.7.0
    # via -r requirements.in
django-entangled==0.6.2
    # via djangocms-frontend
django-filer==3.3.1
    # via
    #   -r requirements.in
    #   djangocms-frontend
django-formtools==2.5.1
    # via django-cms
django-js-asset==3.1.2
    # via django-ckeditor
django-parler==2.3
    # via djangocms-alias
django-polymorphic==4.1.0
    # via django-filer
django-sekizai==4.1.0
    # via django-cms
django-storage-url==0.12.1
    # via -r requirements.in
django-storages==1.14.6
    # via django-storage-url
django-treebeard==4.7.1
    # via django-cms
django-vite==3.1.0
    # via -r requirements.in
djangocms-admin-style==3.3.1
    # via
    #   -r requirements.in
    #   django-cms
djangocms-alias==2.0.3
    # via -r requirements.in
djangocms-attributes-field==4.0.0
    # via
    #   djangocms-frontend
    #   djangocms-link
djangocms-frontend==2.1.2
    # via -r requirements.in
djangocms-link==5.0.1
    # via djangocms-frontend
djangocms-text==0.9.1
    # via -r requirements.in
djangorestframework==3.16.0
    # via -r requirements.in
easy-thumbnails[svg]==2.10
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-frontend
furl==2.1.4
    # via django-storage-url
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.4.0
    # via
    #   djangocms-text
    #   svglib
nh3==0.2.21
    # via djangocms-text
orderedmultidict==1.0.1
    # via furl
packaging==25.0
    # via
    #   build
    #   django-cms
    #   djangocms-text
pillow==11.2.1
    # via
    #   djangocms-text
    #   easy-thumbnails
    #   reportlab
pip-tools==7.4.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via botocore
reportlab==4.4.1
    # via
    #   easy-thumbnails
    #   svglib
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.5.3
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.4.0
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.14.0
    # via dj-database-url
urllib3==2.4.0
    # via botocore
uwsgi==2.0.30
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
wheel==0.45.1
    # via pip-tools
whitenoise==6.9.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
